Rename MockAccountsDb to AccountsDbAsMap
